Switch app.js and index.js
